+++ test/static.test.js
@@ -3,6 +3,7 @@
 // core requires
 var fs = require('fs');
 var path = require('path');
+var net = require('net');
 
 // external requires
 var assert = require('chai').assert;
@@ -32,7 +33,7 @@ describe('static resource plugin', function () {
             log: helper.getLog('server')
         });
 
-        SERVER.listen(PORT, '127.0.0.1', function () {
+        SERVER.listen(0, '127.0.0.1', function () {
             PORT = SERVER.address().port;
             CLIENT = restifyClients.createJsonClient({
                 url: 'http://127.0.0.1:' + PORT,
@@ -100,7 +101,8 @@ describe('static resource plugin', function () {
                     }
 
                     if (testDefault) {
-                        opts.defaultFile = testFileName;
+                        p = '/' + testDir + '/';
+                        opts.default = testFileName;
                         routeName += ' with default';
                     }
                     var re = regex ||
@@ -121,7 +123,61 @@ describe('static resource plugin', function () {
                 });
             });
         });
+    }
 
+    function testNoAppendPath(done, testDefault, tmpDir, regex, staticFile) {
+        var staticContent = '{"content": "abcdefg"}';
+        var staticObj = JSON.parse(staticContent);
+        var testDir = 'public';
+        var testFileName = 'index.json';
+        var routeName = 'GET wildcard';
+        var tmpPath = path.join(__dirname, '../', tmpDir);
+
+        mkdirp(tmpPath, function (err) {
+            assert.ifError(err);
+            DIRS_TO_DELETE.push(tmpPath);
+            var folderPath = path.join(tmpPath, testDir);
+
+            mkdirp(folderPath, function (err2) {
+                assert.ifError(err2);
+
+                DIRS_TO_DELETE.push(folderPath);
+                var file = path.join(folderPath, testFileName);
+
+                fs.writeFile(file, staticContent, function (err3) {
+                    assert.ifError(err3);
+                    FILES_TO_DELETE.push(file);
+                    var p = '/' + testDir + '/' + testFileName;
+                    var opts = { directory: folderPath };
+                    opts.appendRequestPath = false;
+
+                    if (staticFile) {
+                        opts.file = testFileName;
+                    }
+
+                    if (testDefault) {
+                        p = '/' + testDir + '/';
+                        opts.default = testFileName;
+                        routeName += ' with default';
+                    }
+                    var re = regex ||
+                        new RegExp('/' + testDir + '/?.*');
+
+                    SERVER.get({
+                        path: re,
+                        name: routeName
+                    }, plugins.serveStatic(opts));
+
+                    CLIENT.get(p, function (err4, req, res, obj) {
+                        assert.ifError(err4);
+                        assert.equal(res.headers['cache-control'],
+                            'public, max-age=3600');
+                        assert.deepEqual(obj, staticObj);
+                        done();
+                    });
+                });
+            });
+        });
     }
 
     it('static serves static files', function (done) {
@@ -162,4 +218,111 @@ describe('static resource plugin', function () {
         serveStaticTest(done, false, '.tmp', null, true);
     });
 
+    it('static serves static file with appendRequestPath = false',
+        function (done) {
+        testNoAppendPath(done, false, '.tmp');
+    });
+
+    it('static serves default file with appendRequestPath = false',
+        function (done) {
+        testNoAppendPath(done, true, '.tmp');
+    });
+
+    it('restify serve a specific static file with appendRequestPath = false',
+        function (done) {
+        testNoAppendPath(done, false, '.tmp', null, true);
+    });
+
+    // To ensure this will always get properly restored (even in case of a test
+    // failure) we do it here.
+    var originalCreateReadStream = fs.createReadStream;
+    afterEach(function () {
+        fs.createReadStream = originalCreateReadStream;
+    });
+
+    var TMP_PATH = path.join(__dirname, '../', '.tmp');
+    var RAW_REQUEST =
+        'GET /index.html HTTP/1.1\r\n' +
+        'Host: 127.0.0.1:' + PORT + '\r\n' +
+        'User-Agent: curl/7.48.0\r\n' +
+        'Accept: */*\r\n' +
+        '\r\n';
+
+    it('static does not leak the file stream and next() is properly called ' +
+       'when the client disconnects before receiving a reply', function (done) {
+        var streamWasAlreadyCreated = false;
+        var streamWasClosed = false;
+        var socket = new net.Socket();
+
+        fs.createReadStream = function () {
+            // Just in case the code would open more streams in the future.
+            assert.strictEqual(streamWasAlreadyCreated, false);
+            streamWasAlreadyCreated = true;
+
+            var stream = originalCreateReadStream.apply(this, arguments);
+            stream.once('close', function () {
+                streamWasClosed = true;
+            });
+
+            socket.end();
+            return stream;
+        };
+
+        mkdirp(TMP_PATH, function (err) {
+            assert.ifError(err);
+            DIRS_TO_DELETE.push(TMP_PATH);
+            fs.writeFileSync(path.join(TMP_PATH, 'index.html'), 'Hello world!');
+
+            var serve = plugins.serveStatic({
+                directory: TMP_PATH
+            });
+
+            SERVER.get(/.*/, function (req, res, next) {
+                serve(req, res, function (nextRoute) {
+                    assert.strictEqual(streamWasClosed, true);
+                    assert.strictEqual(nextRoute, false);
+                    done();
+                });
+            });
+
+            socket.connect({host: '127.0.0.1', port: PORT}, function () {
+                socket.write(RAW_REQUEST, 'utf-8', function (err2, data) {
+                    assert.ifError(err2);
+                });
+            });
+        });
+    });
+
+    it('static does not open a file stream and next() is properly called ' +
+       'when the client disconnects immediately after sending a request',
+    function (done) {
+        fs.createReadStream = function () {
+            assert(false);
+        };
+
+        mkdirp(TMP_PATH, function (err) {
+            assert.ifError(err);
+            DIRS_TO_DELETE.push(TMP_PATH);
+            fs.writeFileSync(path.join(TMP_PATH, 'index.html'), 'Hello world!');
+
+            var serve = plugins.serveStatic({
+                directory: TMP_PATH
+            });
+
+            SERVER.get(/.*/, function (req, res, next) {
+                serve(req, res, function (nextRoute) {
+                    assert.strictEqual(nextRoute, false);
+                    done();
+                });
+            });
+
+            var socket = new net.Socket();
+            socket.connect({host: '127.0.0.1', port: PORT}, function () {
+                socket.write(RAW_REQUEST, 'utf-8', function (err2, data) {
+                    assert.ifError(err2);
+                    socket.end();
+                });
+            });
+        });
+    });
 });